requirements.txt
requests>=2.32.0
PyYAML>=6.0.2
toml>=0.10.2
python-dotenv>=1.0.1
